Move Most Viewed Pages Macro to livetable #10 * I restructured the packages and replaced the URLBuilder with a proper one.
