+++ application-analytics-api/pom.xml
@@ -24,7 +24,7 @@
   <parent>
     <groupId>com.xwiki.analytics</groupId>
     <artifactId>application-analytics</artifactId>
-    <version>1.6-SNAPSHOT</version>
+    <version>1.0-SNAPSHOT</version>
   </parent>
   <artifactId>application-analytics-api</artifactId>
   <packaging>jar</packaging>
