+++ application-analytics-default/src/main/java/com/xwiki/analytics/internal/configuration/DefaultAnalyticsConfiguration.java
@@ -29,7 +29,7 @@ import org.xwiki.configuration.ConfigurationSource;
 import com.xwiki.analytics.configuration.AnalyticsConfiguration;
 
 /**
- * Default implementation of {@AnalyticsConfiguration}.
+ * Default implementation of {@link AnalyticsConfiguration}.
  *
  * @version $Id$
  */
