+++ application-analytics-ui/pom.xml
@@ -25,7 +25,7 @@
   <parent>
     <groupId>com.xwiki.analytics</groupId>
     <artifactId>application-analytics</artifactId>
-    <version>1.6-SNAPSHOT</version>
+    <version>1.0-SNAPSHOT</version>
   </parent>
   <artifactId>application-analytics-ui</artifactId>
   <packaging>xar</packaging>
